Restricting valid roll interval. Closes #1
HTML attributes for min and max have been added,
which are theoretically used by browsers, but I've
also adjusted the code to take the min and max and
programmatically enforce it on blur from the interval.

Now, while you can go to zero (or below), it will
set up to 1, which is the minimum interval. Technically,
no browsers support an interval of 0 anyways (and really,
an interval of 1 is still not going to hold). The new
setImmediate API in IE10 is interesting, but for now,
I'm sticking with simple timeouts.